webhooks: merge handler_groups.go into groups.go, handler_messages.go into message.go
GroupManagementHandler now lives alongside Group types in groups.go.
Request welcome setters merged into message.go. Removes the artificial
separation between types and their handlers within the same concern.
