+++ webhooks/groups.go
@@ -24,6 +24,7 @@ package webhooks
 
 import (
 	"context"
+	"fmt"
 
 	werrors "github.com/piusalfred/whatsapp/pkg/errors"
 )
@@ -131,4 +132,158 @@ func (handler *Handler) SetGroupStatusUpdateHandler(
 	h GroupStatusUpdateHandler,
 ) {
 	handler.groups.StatusUpdate = h
+}
+// GroupManagementHandler groups all group webhook field handlers into a single
+// dispatch unit. Each field accepts a [MessageChangeValueHandler[Group]] for one
+// WhatsApp group notification type. Leave a field nil to silently skip that
+// notification type (HTTP 200).
+//
+// Group lifecycle events cover create/delete (success and failure variants).
+// Participant events cover joins via invite, join requests, request
+// cancellation, join approval, removal, and departures — each with per-user
+// success/failure arrays. Settings events cover subject, description, and
+// profile picture updates — each with per-field success/failure flags. Status
+// events cover group suspension and clearance.
+//
+// Usage:
+//
+//	gh := &GroupManagementHandler{}
+//	gh.OnLifecycleUpdate(myLifecycleHandler)
+//	gh.OnParticipantsUpdate(myParticipantsHandler)
+type GroupManagementHandler struct {
+	LifecycleUpdate    MessageChangeValueHandler[Group]
+	ParticipantsUpdate MessageChangeValueHandler[Group]
+	SettingsUpdate     MessageChangeValueHandler[Group]
+	StatusUpdate       MessageChangeValueHandler[Group]
+}
+
+// OnLifecycleUpdate sets the handler for group_lifecycle_update webhooks
+// (group creation and deletion, with success and failure variants).
+func (gh *GroupManagementHandler) OnLifecycleUpdate(
+	fn func(ctx context.Context, nctx *MessageNotificationContext, groups []*Group) error,
+) {
+	gh.LifecycleUpdate = MessageChangeValueHandlerFunc[Group](fn)
+}
+
+// OnParticipantsUpdate sets the handler for group_participants_update webhooks
+// (participants joining via invite, requesting to join, cancelling requests,
+// join request approval, participant removal, and participant departures).
+func (gh *GroupManagementHandler) OnParticipantsUpdate(
+	fn func(ctx context.Context, nctx *MessageNotificationContext, groups []*Group) error,
+) {
+	gh.ParticipantsUpdate = MessageChangeValueHandlerFunc[Group](fn)
+}
+
+// OnSettingsUpdate sets the handler for group_settings_update webhooks
+// (group subject, description, and profile picture changes with per-field
+// success/failure reporting).
+func (gh *GroupManagementHandler) OnSettingsUpdate(
+	fn func(ctx context.Context, nctx *MessageNotificationContext, groups []*Group) error,
+) {
+	gh.SettingsUpdate = MessageChangeValueHandlerFunc[Group](fn)
+}
+
+// OnStatusUpdate sets the handler for group_status_update webhooks
+// (group suspension and suspension clearance).
+func (gh *GroupManagementHandler) OnStatusUpdate(
+	fn func(ctx context.Context, nctx *MessageNotificationContext, groups []*Group) error,
+) {
+	gh.StatusUpdate = MessageChangeValueHandlerFunc[Group](fn)
+}
+
+// IsGroupManagementWebhook reports whether field is one of the four group
+// management webhook fields (lifecycle, participants, settings, status).
+func (f ChangeField) IsGroupManagementWebhook() bool {
+	switch f {
+	case ChangeFieldGroupLifecycleUpdate,
+		ChangeFieldGroupParticipantsUpdate,
+		ChangeFieldGroupSettingsUpdate,
+		ChangeFieldGroupStatusUpdate:
+		return true
+	default:
+		return false
+	}
+}
+
+// GroupChangeFields returns every [ChangeField] that triggers a group
+// management webhook notification.
+func GroupChangeFields() []ChangeField {
+	return []ChangeField{
+		ChangeFieldGroupLifecycleUpdate,
+		ChangeFieldGroupParticipantsUpdate,
+		ChangeFieldGroupSettingsUpdate,
+		ChangeFieldGroupStatusUpdate,
+	}
+}
+
+// Handle dispatches the group notification to the correct handler based on
+// change.Field. Nil handlers are silently skipped (HTTP 200). Errors from user
+// handlers are routed through onError; if onError returns a non-nil error,
+// processing stops.
+//
+//nolint:gocognit // dispatch switch
+func (gh *GroupManagementHandler) Handle(
+	ctx context.Context,
+	change Change,
+	entry Entry,
+	onError func(ctx context.Context, err error) error,
+) error {
+	nctx := &MessageNotificationContext{
+		EntryID:          entry.ID,
+		MessagingProduct: change.Value.MessagingProduct,
+		Metadata:         change.Value.Metadata,
+		Contacts:         change.Value.Contacts,
+	}
+
+	switch change.Field {
+	case ChangeFieldGroupLifecycleUpdate.String():
+		if gh.LifecycleUpdate == nil {
+			return nil
+		}
+		if err := gh.LifecycleUpdate.Handle(ctx, nctx, change.Value.Groups); err != nil {
+			if onError != nil {
+				if handlerErr := onError(ctx, err); handlerErr != nil {
+					return fmt.Errorf("error handler: %w", handlerErr)
+				}
+			}
+		}
+
+	case ChangeFieldGroupParticipantsUpdate.String():
+		if gh.ParticipantsUpdate == nil {
+			return nil
+		}
+		if err := gh.ParticipantsUpdate.Handle(ctx, nctx, change.Value.Groups); err != nil {
+			if onError != nil {
+				if handlerErr := onError(ctx, err); handlerErr != nil {
+					return fmt.Errorf("error handler: %w", handlerErr)
+				}
+			}
+		}
+
+	case ChangeFieldGroupSettingsUpdate.String():
+		if gh.SettingsUpdate == nil {
+			return nil
+		}
+		if err := gh.SettingsUpdate.Handle(ctx, nctx, change.Value.Groups); err != nil {
+			if onError != nil {
+				if handlerErr := onError(ctx, err); handlerErr != nil {
+					return fmt.Errorf("error handler: %w", handlerErr)
+				}
+			}
+		}
+
+	case ChangeFieldGroupStatusUpdate.String():
+		if gh.StatusUpdate == nil {
+			return nil
+		}
+		if err := gh.StatusUpdate.Handle(ctx, nctx, change.Value.Groups); err != nil {
+			if onError != nil {
+				if handlerErr := onError(ctx, err); handlerErr != nil {
+					return fmt.Errorf("error handler: %w", handlerErr)
+				}
+			}
+		}
+	}
+
+	return nil
 }
