--- webhooks/handler_messages.go
@@ -1,39 +1,0 @@
-//  Copyright 2023 Pius Alfred <me.pius1102@gmail.com>
-//
-//  Permission is hereby granted, free of charge, to any person obtaining a copy of this software
-//  and associated documentation files (the "Software"), to deal in the Software without restriction,
-//  including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense,
-//  and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so,
-//  subject to the following conditions:
-//
-//  The above copyright notice and this permission notice shall be included in all copies or substantial
-//  portions of the Software.
-//
-//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT
-//  LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
-//  IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
-//  WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
-//  SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-
-// Request welcome message handler registration for Handler.
-
-package webhooks
-
-import (
-	"context"
-)
-
-// OnRequestWelcomeMessage registers a handler for request_welcome messages in the messages webhook.
-func (handler *Handler) OnRequestWelcomeMessage(
-	fn func(ctx context.Context, notificationCtx *MessageNotificationContext,
-		info *MessageInfo, media *Message) error,
-) {
-	handler.messages.RequestWelcome = MessageHandlerFunc[Message](fn)
-}
-
-// SetRequestWelcomeMessageHandler sets the handler for request_welcome messages.
-func (handler *Handler) SetRequestWelcomeMessageHandler(
-	fn RequestWelcomeMessageHandler,
-) {
-	handler.messages.RequestWelcome = fn
-}
